requirements.txt
streamlit
crewai
transformers
torch
faiss-cpu
PyPDF2
requests
google-generativeai
litellm
youtube-transcript-api
document-processor
